Refactored the models to comply with Robert's working model. It should work now.
